requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deep-translator==1.11.4
distro==1.9.0
dotenv==0.9.9
easyocr==1.7.2
exceptiongroup==1.3.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.4.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-genai==1.36.0
google-pasta==0.2.0
grpcio==1.74.0
gTTS==2.5.4
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jpype1==1.6.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
konlpy==0.6.0
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.1.0
narwhals==2.3.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.107.2
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pip==25.2
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.13.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.2
ydf==0.8.0
